Ajustes para firebase
