+++ src/components/Register.js
@@ -22,13 +22,40 @@ export const Register = () => {
   // Se crea contenedor para inicio de sesión con correo, google y facebook
   const section1 = document.createElement('section');
   section1.className = 'section1';
+
+  const nameAndLastNameContainer = document.createElement('section');
+  nameAndLastNameContainer.className = 'name-lastName-container';
+
+  const inputName = document.createElement('input');
+  inputName.setAttribute('requiered', '');
+  inputName.setAttribute('placeholder', 'Nombre');
+  inputName.setAttribute('id', 'name');
+
+  const inputLastName = document.createElement('input');
+  inputLastName.setAttribute('requiered', '');
+  inputLastName.setAttribute('placeholder', 'Apellido');
+  inputLastName.setAttribute('id', 'lastName');
+
   const inputEmail = document.createElement('input');
   inputEmail.setAttribute('requiered', '');
   inputEmail.setAttribute('placeholder', 'Correo electrónico');
   inputEmail.setAttribute('id', 'emailLogin');
 
+  const inputPassword = document.createElement('input');
+  inputPassword.setAttribute('requiered', '');
+  inputPassword.setAttribute('placeholder', 'Crear contraseña');
+  inputPassword.setAttribute('id', 'password');
+
+  const inputConfirmPassword = document.createElement('input');
+  inputConfirmPassword.setAttribute('requiered', '');
+  inputConfirmPassword.setAttribute('placeholder', 'Confirma tu contraseña');
+  inputConfirmPassword.setAttribute('id', 'confirm-password');
+
+  nameAndLastNameContainer.append(inputName, inputLastName);
+
   const buttonSignUp = document.createElement('button');
-  buttonSignUp.textContent = 'Registrarse';
+  buttonSignUp.textContent = 'Regístrate';
+
   const google = document.createElement('a');
   google.setAttribute('href', '');
   google.textContent = 'Registrarse con google';
@@ -37,24 +64,33 @@ export const Register = () => {
   facebook.setAttribute('href', '');
   facebook.textContent = 'Registrarse con facebook';
   facebook.className = 'facebook';
-  section1.append(inputEmail, buttonSignUp, google, facebook);
+  section1.append(
+    inputEmail,
+    nameAndLastNameContainer,
+    inputEmail,
+    inputPassword,
+    inputConfirmPassword,
+    buttonSignUp,
+    google,
+    facebook,
+  );
 
-  // Se crea contendeor pra registrarse si no tien una cuenta
+  // Se crea contendeor pra registrarse si no tiene una cuenta
   const section2 = document.createElement('section');
   section2.className = 'section2';
   const account = document.createElement('p');
   account.textContent = '¿Tienes Cuenta?';
   account.className = 'account';
-  const linkRegister = document.createElement('a');
-  linkRegister.setAttribute('href', '');
-  linkRegister.textContent = ' Inicia Sesión';
-  section2.append(account, linkRegister);
+  const linkLogin = document.createElement('a');
+  linkLogin.setAttribute('href', '');
+  linkLogin.textContent = ' Inicia Sesión';
+  section2.append(account, linkLogin);
 
   buttonSignUp.addEventListener('click', () => {
-    onNavigate('/register2');
-  });
-  linkRegister.addEventListener('click', () => {
     onNavigate('/wall');
+  });
+  linkLogin.addEventListener('click', () => {
+    onNavigate('/');
   });
 
   div.append(header, section1, section2);
