+++ src/components/Wall.js
@@ -4,26 +4,23 @@ export const Wall = () => {
   // contenedor que almacenará los 2 botones y dará un solo return
   const div = document.createElement('div');
   // se crea titulo y parrafo
-  const title = document.createElement('h1');
-  const paragraph = document.createElement('h2');
-  // Se crean 2 botones
-  const button = document.createElement('button');
-  const buttonBack = document.createElement('button');
-  const inputEmail = document.createElement('input');
-  const inputPass = document.createElement('input');
+  const header = document.createElement('header');
+  header.className = 'header';
 
-  paragraph.textContent = 'Inicia Sesión';
-  button.textContent = 'Entrar';
-  buttonBack.textContent = 'Regresar';
+  const title = document.createElement('h1');
+  title.className = 'title';
   title.textContent = 'Yummy Friends';
-  button.addEventListener('click', () => {
-    onNavigate('/');
-  });
+
+  const buttonBack = document.createElement('button');
+  buttonBack.textContent = 'Cerrar sesión';
+  buttonBack.className = 'buttonBack';
+
+  header.append(title, buttonBack);
   buttonBack.addEventListener('click', () => {
     onNavigate('/');
   });
 
-  div.append(title, paragraph, inputEmail, inputPass, button, buttonBack);
+  div.append(header);
 
   return div;
 };
