+++ src/main.js
@@ -1,8 +1,6 @@
 import { Welcome } from './components/Welcome.js';
-import { Wall } from './components/Wall.js';
-
 import { Register } from './components/Register.js';
-import { Register2 } from './components/Register2.js';
+import { Wall } from './components/Wall.js';
 
 // variable que acceda al nodo en HTML llamado root
 const root = document.getElementById('root');
@@ -10,9 +8,9 @@ const root = document.getElementById('root');
 // Variable que contendra un objeto que contiene la ruta y o que debe renderizar
 const routes = {
   '/': Welcome, // cuando entre en '/' debe renderizar la página inicial
-  '/wall': Wall,
   '/register': Register,
-  '/register2': Register2,
+  '/wall': Wall,
+
 };
 // recibe la ruta, despues utiliza el objeto window.history y lanza el pushState
 export const onNavigate = (pathname) => {
