+++ src/style.css
@@ -22,16 +22,19 @@ img {
 }
 
 a {
-    font-size: 12px;
+    font-size: 14px;
+    background-color: #ffffff;
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;   
+    border-radius: 5px; 
 }
 
 button {
     margin: 10px;
     justify-content: center;
-    border-radius: 10px;
+    border-radius: 5px;
     padding: 5px;
     width: 118px;
-    background-color: #8B6767;
+    background-color: #A3445D;
     color: #FFFFFF;
     font-family: 'Lexend Giga', sans-serif;
     font-size: 12px;
@@ -73,10 +76,12 @@ button {
     display: flex;
     text-align: center;
     flex-direction: column;
-    background-color: #ffffff;
+    background-color: #C8BBBB;
     margin: 30px 30px;
     padding: 10px;
-    align-items: center;
+    align-items: center;    
+    box-shadow: 2px 5px 5px  rgba(129, 122, 124, 0.5) inset;
+    border-radius: 10px;
 }
 
 .section1 button {
@@ -87,7 +92,7 @@ button {
     margin-top: 10px;
     width: 100%;
     padding: 6px 0px;
-    background-color: #D9D9D9;
+    background-color: #ffffff;
     border: none;
 }
 .section1 input::placeholder{
@@ -98,19 +103,19 @@ button {
     margin-top: 10px;
     width: 100%;
     padding: 6px 0px;
-    background-color: #D9D9D9;
+    background-color: #ffffff;
     border: none;
 }
 
 .google {
     padding: 5px 14px;
-    background-color: #D9D9D9;
+    background-color: #ffffff;
 }
 
 .facebook {
     margin-top: 10px;
     padding: 5px 4px;
-    background-color: #D9D9D9;
+    background-color: #ffffff;
 }
 
 .facebookImage {
@@ -122,15 +127,17 @@ button {
 .section2 {
     display: flex;
     align-content: center;
-    background-color: #ffffff;
+    background-color: #C8BBBB;
     margin: 0px 29px;
-    padding: 13px 5px;
+    padding: 13px;
     justify-content: space-between;
-    border: solid 1px;
+    box-shadow: 2px 5px 5px  rgba(129, 122, 124, 0.5) inset;
+    border-radius: 10px;
 }
 
 .account {
     font-size: 12px;
+    background-color: #ffffff;
 }
 
 
@@ -148,7 +155,7 @@ button {
     margin-top: 10px;
     width: 100%;
     padding: 6px 0px;
-    background-color: #D9D9D9;
+    background-color: #ffffff;
     border: none;
 }
 
@@ -167,4 +174,12 @@ button {
 }
 .section2-Register a {
     font-size: 14px;
+}
+
+/*Estilos de la sección Muro */
+
+.buttonBack{
+    background-color: #5A51C5;
+    font-size: 12px;
+    width: 50%;
 }
